+++ packages/picmo/src/views/CategoryTabs.template.ts
@@ -1,7 +1,9 @@
 import { Template } from '../Template';
 
 export default new Template(({ classes }) => /* html */`
-  <ul role="tablist" class="${classes.categoryButtons}">
-    <div data-placeholder="tabs"></div>
-  </ul>
+  <div class="${classes.categoryButtonsContainer}">
+    <ul role="tablist" class="${classes.categoryButtons}">
+      <div data-placeholder="tabs"></div>
+    </ul>
+  </div>
 `);