+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 For full details, see the documentation at [https://picmojs.com](https://picmojs.com).
 
+## [5.6.0] - TBD
+
+[`picmo`] Added ability to update certain picker options after creation.
+[`@picmo/popup-picker`] Added ability to specify new reference and trigger elements when opening the picker.
+
 ## [5.5.2] - 2022-08-17
 - Fixed `process.env` appearing in build output.